+++ packages/web/src/app/search/page.tsx
@@ -5,25 +5,27 @@ import {
     ResizablePanel,
     ResizablePanelGroup,
 } from "@/components/ui/resizable";
+import { ScrollArea } from "@/components/ui/scroll-area";
 import { Separator } from "@/components/ui/separator";
+import useCaptureEvent from "@/hooks/useCaptureEvent";
 import { useNonEmptyQueryParam } from "@/hooks/useNonEmptyQueryParam";
+import { SearchQueryParams, SearchResultFile } from "@/lib/types";
 import { createPathWithQueryParams } from "@/lib/utils";
 import { SymbolIcon } from "@radix-ui/react-icons";
+import { Scrollbar } from "@radix-ui/react-scroll-area";
 import { useQuery } from "@tanstack/react-query";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
-import { useCallback, useEffect, useMemo, useState } from "react";
+import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import logoDark from "../../../public/sb_logo_dark.png";
 import logoLight from "../../../public/sb_logo_light.png";
 import { search } from "../api/(client)/client";
 import { SearchBar } from "../searchBar";
 import { SettingsDropdown } from "../settingsDropdown";
-import useCaptureEvent from "@/hooks/useCaptureEvent";
 import { CodePreviewPanel } from "./components/codePreviewPanel";
+import { FilterPanel } from "./components/filterPanel";
 import { SearchResultsPanel } from "./components/searchResultsPanel";
-import { SearchQueryParams, SearchResultFile } from "@/lib/types";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { Scrollbar } from "@radix-ui/react-scroll-area";
+import { ImperativePanelHandle } from "react-resizable-panels";
 
 const DEFAULT_MAX_MATCH_DISPLAY_COUNT = 200;
 
@@ -32,9 +34,6 @@ export default function SearchPage() {
     const searchQuery = useNonEmptyQueryParam(SearchQueryParams.query) ?? "";
     const _maxMatchDisplayCount = parseInt(useNonEmptyQueryParam(SearchQueryParams.maxMatchDisplayCount) ?? `${DEFAULT_MAX_MATCH_DISPLAY_COUNT}`);
     const maxMatchDisplayCount = isNaN(_maxMatchDisplayCount) ? DEFAULT_MAX_MATCH_DISPLAY_COUNT : _maxMatchDisplayCount;
-
-    const [selectedMatchIndex, setSelectedMatchIndex] = useState(0);
-    const [selectedFile, setSelectedFile] = useState<SearchResultFile | undefined>(undefined);
 
     const captureEvent = useCaptureEvent();
 
@@ -151,80 +150,148 @@ export default function SearchPage() {
                     />
                 </div>
                 <Separator />
-                <div className="bg-accent py-1 px-2 flex flex-row items-center gap-4">
-                    {
-                        isLoading ? (
-                            <p className="text-sm font-medium">Loading...</p>
-                        ) : fileMatches.length > 0 && searchResponse ? (
-                            <p className="text-sm font-medium">{`[${searchDurationMs} ms] Displaying ${numMatches} of ${searchResponse.Result.MatchCount} matches in ${fileMatches.length} ${fileMatches.length > 1 ? 'files' : 'file'}`}</p>
-                        ) : (
-                            <p className="text-sm font-medium">No results</p>
-                        )
-                    }
-                    {isMoreResultsButtonVisible && !isLoading && (
-                        <div
-                            className="cursor-pointer text-blue-500 text-sm hover:underline"
-                            onClick={onLoadMoreResults}
-                        >
-                            (load more)
-                        </div>
-                    )}
-                </div>
+                {!isLoading && (
+                    <div className="bg-accent py-1 px-2 flex flex-row items-center gap-4">
+                        {
+                            fileMatches.length > 0 && searchResponse ? (
+                                <p className="text-sm font-medium">{`[${searchDurationMs} ms] Displaying ${numMatches} of ${searchResponse.Result.MatchCount} matches in ${fileMatches.length} ${fileMatches.length > 1 ? 'files' : 'file'}`}</p>
+                            ) : (
+                                <p className="text-sm font-medium">No results</p>
+                            )
+                        }
+                        {isMoreResultsButtonVisible && (
+                            <div
+                                className="cursor-pointer text-blue-500 text-sm hover:underline"
+                                onClick={onLoadMoreResults}
+                            >
+                                (load more)
+                            </div>
+                        )}
+                    </div>
+                )}
                 <Separator />
             </div>
 
-            {/* Search Results & Code Preview */}
-            <ResizablePanelGroup direction="horizontal">
-                <ResizablePanel minSize={20}>
-                    {isLoading ? (
-                        <div className="flex flex-col items-center justify-center h-full gap-2">
-                            <SymbolIcon className="h-6 w-6 animate-spin" />
-                            <p className="font-semibold text-center">Searching...</p>
-                        </div>
-                    ) : fileMatches.length > 0 ? (
-                        <ScrollArea
-                            className="h-full"
-                        >
-                            <SearchResultsPanel
-                                fileMatches={fileMatches}
-                                onOpenFileMatch={(fileMatch) => {
-                                    setSelectedFile(fileMatch);
-                                }}
-                                onMatchIndexChanged={(matchIndex) => {
-                                    setSelectedMatchIndex(matchIndex);
-                                }}
-                            />
-                            {isMoreResultsButtonVisible && (
-                                <div className="p-3">
-                                    <span
-                                        className="cursor-pointer text-blue-500 hover:underline"
-                                        onClick={onLoadMoreResults}
-                                    >
-                                        Load more results
-                                    </span>
-                                </div>
-                            )}
-                            <Scrollbar orientation="vertical" />
-                        </ScrollArea>
-                    ) : (
-                        <div className="flex flex-col items-center justify-center h-full">
-                            <p className="text-sm text-muted-foreground">No results found</p>
-                        </div>
-                    )}
-                </ResizablePanel>
-                <ResizableHandle withHandle={selectedFile !== undefined} />
-                <ResizablePanel
-                    minSize={20}
-                    hidden={!selectedFile}
-                >
-                    <CodePreviewPanel
-                        fileMatch={selectedFile}
-                        onClose={() => setSelectedFile(undefined)}
-                        selectedMatchIndex={selectedMatchIndex}
-                        onSelectedMatchIndexChange={setSelectedMatchIndex}
-                    />
-                </ResizablePanel>
-            </ResizablePanelGroup>
+            {isLoading ? (
+                <div className="flex flex-col items-center justify-center h-full gap-2">
+                    <SymbolIcon className="h-6 w-6 animate-spin" />
+                    <p className="font-semibold text-center">Searching...</p>
+                </div>
+            ) : (
+                <PanelGroup
+                    fileMatches={fileMatches}
+                    isMoreResultsButtonVisible={isMoreResultsButtonVisible}
+                    onLoadMoreResults={onLoadMoreResults}
+                />
+            )}
         </div>
     );
+}
+
+interface PanelGroupProps {
+    fileMatches: SearchResultFile[];
+    isMoreResultsButtonVisible?: boolean;
+    onLoadMoreResults: () => void;
+}
+
+const PanelGroup = ({
+    fileMatches,
+    isMoreResultsButtonVisible,
+    onLoadMoreResults,
+}: PanelGroupProps) => {
+    const [selectedMatchIndex, setSelectedMatchIndex] = useState(0);
+    const [selectedFile, setSelectedFile] = useState<SearchResultFile | undefined>(undefined);
+    const [filteredFileMatches, setFilteredFileMatches] = useState<SearchResultFile[]>(fileMatches);
+
+    const codePreviewPanelRef = useRef<ImperativePanelHandle>(null);
+    useEffect(() => {
+        if (selectedFile) {
+            codePreviewPanelRef.current?.expand();
+        } else {
+            codePreviewPanelRef.current?.collapse();
+        }
+    }, [selectedFile]);
+
+    return (
+        <ResizablePanelGroup
+            direction="horizontal"
+        >
+            {/* ~~ Filter panel ~~ */}
+            <ResizablePanel
+                minSize={20}
+                maxSize={30}
+                defaultSize={20}
+                collapsible={true}
+                id={'filter-panel'}
+                order={1}
+            >
+                <FilterPanel
+                    matches={fileMatches}
+                    onFilterChanged={(filteredFileMatches) => {
+                        setFilteredFileMatches(filteredFileMatches)
+                    }}
+                />
+            </ResizablePanel>
+            <ResizableHandle
+                className="bg-accent w-1 transition-colors delay-50 data-[resize-handle-state=drag]:bg-accent-foreground data-[resize-handle-state=hover]:bg-accent-foreground"
+            />
+
+            {/* ~~ Search results ~~ */}
+            <ResizablePanel
+                minSize={10}
+                id={'search-results-panel'}
+                order={2}
+            >
+                {filteredFileMatches.length > 0 ? (
+                    <ScrollArea
+                        className="h-full"
+                    >
+                        <SearchResultsPanel
+                            fileMatches={filteredFileMatches}
+                            onOpenFileMatch={(fileMatch) => {
+                                setSelectedFile(fileMatch);
+                            }}
+                            onMatchIndexChanged={(matchIndex) => {
+                                setSelectedMatchIndex(matchIndex);
+                            }}
+                        />
+                        {isMoreResultsButtonVisible && (
+                            <div className="p-3">
+                                <span
+                                    className="cursor-pointer text-blue-500 hover:underline"
+                                    onClick={onLoadMoreResults}
+                                >
+                                    Load more results
+                                </span>
+                            </div>
+                        )}
+                        <Scrollbar orientation="vertical" />
+                    </ScrollArea>
+                ) : (
+                    <div className="flex flex-col items-center justify-center h-full">
+                        <p className="text-sm text-muted-foreground">No results found</p>
+                    </div>
+                )}
+            </ResizablePanel>
+            <ResizableHandle
+                withHandle={selectedFile !== undefined}
+            />
+
+            {/* ~~ Code preview ~~ */}
+            <ResizablePanel
+                ref={codePreviewPanelRef}
+                minSize={10}
+                collapsible={true}
+                id={'code-preview-panel'}
+                order={3}
+            >
+                <CodePreviewPanel
+                    fileMatch={selectedFile}
+                    onClose={() => setSelectedFile(undefined)}
+                    selectedMatchIndex={selectedMatchIndex}
+                    onSelectedMatchIndexChange={setSelectedMatchIndex}
+                />
+            </ResizablePanel>
+        </ResizablePanelGroup>
+    )
 }